pyproject.toml
[project]
name = "mujoco_ravens"
version = "0.0.0"
description = "Simulation environments for basic manipulation tasks."
authors = [
	{ name = "Peter David Fagan", email = "peterdavidfagan@gmail.com" }
]
readme = "README.md"

[tool.poetry]
name = "mujoco_ravens"
version = "0.0.0"
description = "Simulation environments for basic manipulation tasks."
license = "Apache-2.0"
authors = [
	"Peter David Fagan <peterdavidfagan@gmail.com>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.6"
mujoco = "^2.3.0"
dm-control = "1.0.9"
dm_robotics-moma = "0.5.0"
dm_robotics-transformations = "0.5.0"
dm_robotics-agentflow = "0.5.0"
dm_robotics-geometry = "0.5.0"
dm_robotics-manipulation = "0.5.0"
dm_robotics-controllers = "0.5.0"

pillow = "10.0.0"
jax = "^0.4.14"
jaxlib = "^0.4.14"
matplotlib = "^3.7.2"
hydra-core = "^1.3.2"

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120
